use name attribute over unavailable fqdn
Foreman delegates the fqdn to the primary interface. When a host is
deleted, the primary interface gets deleted first and the fqdn is not
available in the before_destroy callback. The host's name attribute
however is syncronized with the primary nic's name attribute.
The method Nic::Interface#normalize_name ensures that name is always a
fqdn.