Use the parsed virtual fact provided by 1.24
Foreman 1.24 introduced an extended the reported data facets with
whether a host is virtual. This is more reliable than the provider since
not everyone is using the compute resource integration. That is still
present as a fallback since it may not be known in various cases.  For
example, when provisioning or if no fact importing happens.